fixed warning

+++ src/utilities.cpp
@@ -9,7 +9,7 @@
 size_t TO_REF(int64_t v) {
 #if !defined(NDEBUG) && defined(_DEBUG)
 	uint64_t m = (size_t)(-1);
-	assert(v <= m);
+	assert((uint64_t)v <= m);
 #endif
 	return (size_t)v;
 }
